Bug 1468943 - Address review requests
